Split validate command into 2.
The new one is CI specific to make things a bit more clear.
This is a first step in cleaning up the error message when values are not
provided. Right now it's not documented anywhere which arguments are required
for which command/mode.

Also add a bit more checks on args for the different subcommand in order
to make it more clear to users which args should be used (or not) for
each command.

Bug: N/A
Test: N/A
Change-Id: I78f5345315ec57606d2e364c21d4b324e70d1260