Fix validate command.
wow, nothing really worked. Tons of issues everywhere.

Ended up changing how we find the max AGP, and also removed
the hardcoded map from AGP to Gradle, and use version_mappings.txt
for both.

Since we need to read that file (and others like the gradle wrapper
files), I'm now finding the root folder of the git project, and passing
that along to find what's needed.

Also realized that we need to have an Android SDK to run these projects
so for now do a simple check on the ANDROID_HOME env var. We can't
really create local.properties as validation uses the RELEASE
conversion mode which is the same as the publish. To create
local.properties we'd have to create a separate PUBLISH mode that
does not create the local file.

There are a few recipes that while they could possibly work
on AGP 8.0, they are problematic because of the bytecode
target and how we can make it targe a specific version (17,
same as Kotlin, which is picked up from the runtime JDK).
We could set the bytecode level in the build file but the
pattern for that changed in 8.1, so instead of having to
make that part of the recipe dynamic, we just move them
to 8.1. The state of the new API is much better in 8.1+
anyway.

Bug: N/A
Test: N/A
Change-Id: I453d5f6685d45db60a625a436256a1e4be379895