Make Kotlin version per AGP
This includes reading the kotlin version from version_mappings.txt
and some refactoring here and there to make reading it and using it
simpler.

Also added a 3rd AGP to test (8.2), and some work to make older
AGP and older Kotlin versions available as targets

Still need to figure out the target issue in the bazel macro.

Bug: N/A
Test: tested via the recipes themselves
Change-Id: Ibf9029291d54ebd66e1988dce9c95fdf6fab049f