Allow for separate recipe name and location
Replaced the previously unused 'title' entry in the toml file with
'indexName' that is used as display name in the index. Also added
'destinationFolder'.

The goal is to allow us to have the "same" recipe name, and link,
while having a different source folder. This will be useful once we
start forking recipe to have different versions compatible with
different major versions of AGP.

This also opens the door with having display names in the index
be different from the folder name, since camel case isn't super
pretty.

Bug: N/A
Test: N/A
Change-Id: I73388385d21ab2f030c101fa30902e69d675be84