Fix destination overwrite issue
A previous change made the destination of the converter be the
parent folder of the recipe folder rather thant the folder itself,
leaving the decision of the folder name to the recipe (via metadata).

This change also impacted detection of existing destination folder, and
while it worked for batch conversion, it did not for single conversion.

This change fixes this, by separating the logic for single vs batch
conversion.

Single conversion expect the recipe folder to be empty, but the
parent (which is passed to the converted) can contain other things.
Overwritting will not impact the parent but the recipe folder only.

Batch conversion expect the recipe folder parent to be empty since
it's going to receive a lot of recipes. Overwritting will delete
the whole content of the parent (--destination) folder.

This also replaces all error handling to use System.exit instead
of exceptions as it's more user friendly (there's an option to
throw for debugging if needed)

Bug: N/A
Test: existing
Change-Id: Ie115ce6b8b13e4c77c0f99371f0d37d89e2f7a7b