Fix workflow to deal with different jar location
The convert-tool build is done via a different build command
to run a single jar. Because the jar location is used to find the root
of the git project inside the tool (in order to find files needed for
the recipe conversion), this fails to compute the right folder.

This change introduces an env var that triggers a different computation
(since it's only needed for the github workflow, it was better than
adding a command line args -- also it seems like the parser API only
supports (key, value) pair type of arguments)

Bug: N/A
Test: N/A
Change-Id: I12b33be19dc92f04a17a14ec70d93bf4e5db5757