Improve recipes.bzl
Previously, the "ToT" targets depended on all versions of Gradle listed
in the version_mappings.txt file. The reason for this is because the
"ToT" targets checked that the versions listed in version_mappings.txt
were the same versions being tested.

With this change, we still check that the versions listed in
version_mappings.txt are the same versions being tested, but it's done
in a way that doesn't require different versions of Gradle being
dependencies of the "ToT" targets.

Bug: NA
Test: Existing
Change-Id: I4cc64da1cd21f88f741b7061cffaaf1a79632fb8