Added recipe to show bridging from legacy Tasks.
When using legacy tasks like Gradle's Copy or JavaCompile, inputs
and outputs are usually expressed using File instance instead of
Property instances. This can be problematic with the Variant API
which requires tasks to express their inputs/outputs using Property
instances.

There are probably various ways to bridge the two correctly,
added an example on how to bridge the output (which is the most
likely usage in the Variant API).

Test: This is a test.
Bug: N/A
Change-Id: I1ea759fe5830b03053f0c7bd7fe61819880018d4