New recipe for toListenTo API
This recipe shows a basic example of toListenTo on the APK
artifact. However, because it needs to be validated with a task
that doesn not depend on the creation (rendering the toListenTo
setup moot), this also introduces a new validationTask that runs
after normal tasks.

This change also fix a series of issues in convert-tool:
- (somewhat) recent change to destination folder handling was not
  supported properly in the validation task
- recent changes to AGP versions in version_mapping didn't update
  the tool fully and this needed correction. In particular we also
  need to download the maven listing file to find the right AGP version.

In order the facilitate the above and make code easier to understand
with regards to fully AGP version (ie published version) and short
versions (major.minor only), this change also introduces some
new type to manipulate full and short AGP versions, so that they
are type safe and cannot be mixed and matched.

Bug: N/A
Test: existing tests, though we need to introduce tool testing in presubmit.
Change-Id: I54b7582f2187e62abdfc4d4be6e1c386b928137f