New Context class to replace global vars
There's been a proliferation of global variables and methods
that makes things problematic for testing on presubmit where
we cannot do some things (like downloading from gmaven).

This CL (re)introduces a context objects that has to be passed
to all the converters and validator (we used to have this just
for branchRoot, but now we have an object that provides that
info and much more).

This change also split the parsing of VersionMapping and the
processing of maven-medata.xml so that actions that only need
the former don't need to do the latter, and therefore can run
unchanged on presubmit.

Bug: N/A
Test: existing tests. Next up is integration test of the convert-tool on our presubmit
Change-Id: If7483d21fc9c49c4b28f21c850431e3e530fec20