Use ConvertTool's CLI for gradle-recipe tests
Previously, ConvertTool's CLI didn't have any test coverage. With this
change, GradleRecipeTest mimics the developer workflow described in
CONTRIBUTING.md.

This new test coverage uncovered a bug in the convert tool when run on
Windows; it would always fail because of a call to
Files::getPosixFilePermissions in Converter.kt. This change fixes that
bug by avoiding that call on Windows.

This change also removes the convert tool's validateCI sub-command and
adds a --ci flag to the validate sub-command instead (in order to keep
the same code path for tests and the developer workflow in
CONTRIBUTING.md).

Bug: NA
Test: GradleRecipeTest
Change-Id: I438d4be195432ff1a9ef27cde27ba4d1f816e71c